+++ src/fourthLesson/shape/TestArea.java
@@ -1,0 +1,14 @@
+package fourthLesson.shape;
+
+public class TestArea  {
+
+
+    public static void main(String[] args) {
+        Area area = new Area();
+        area.circleArea(5);
+        area.rectangleArea(6,3);
+        area.squareArea(4);
+
+
+    }
+}
